Moved the PlayerBuildeTest
Now under the correct package.
